chore :  예외처리 추가, 오타수정

+++ src/main/java/org/ezcode/codetest/domain/game/service/GameManagementDomainService.java
@@ -2,6 +2,8 @@ package org.ezcode.codetest.domain.game.service;
 
 import java.util.List;
 
+import org.ezcode.codetest.domain.game.exception.GameException;
+import org.ezcode.codetest.domain.game.exception.GameExceptionCode;
 import org.ezcode.codetest.domain.game.model.entity.Item;
 import org.ezcode.codetest.domain.game.model.entity.Skill;
 import org.ezcode.codetest.domain.game.repository.ItemRepository;
@@ -17,13 +19,20 @@ public class GameManagementDomainService {
 	private final ItemRepository itemRepository;
 	private final SkillRepository skillRepository;
 
-
 	public Item createItem(Item item) {
+
+		boolean exists = itemRepository.existsByName(item.getName());
+
+		if(exists) throw new GameException(GameExceptionCode.ITEM_ALREADY_EXISTS);
 
 		return itemRepository.save(item);
 	}
 
 	public void deleteItem(String name) {
+
+		boolean exists = itemRepository.existsByName(name);
+
+		if(!exists) throw new GameException(GameExceptionCode.ITEM_NOT_FOUND);
 
 		itemRepository.deleteByName(name);
 	}
@@ -35,10 +44,18 @@ public class GameManagementDomainService {
 
 	public Skill createSkill(Skill skill) {
 
+		boolean exists = skillRepository.existsByName(skill.getName());
+
+		if(exists) throw new GameException(GameExceptionCode.SKILL_ALREADY_EXISTS);
+
 		return skillRepository.save(skill);
 	}
 
 	public void deleteSkill(String name) {
+
+		boolean exists = skillRepository.existsByName(name);
+
+		if(!exists) throw new GameException(GameExceptionCode.SKILL_NOT_FOUND);
 
 		skillRepository.deleteByName(name);
 	}
